requirements.txt
paddlepaddle-gpu==2.1.0
opencv-python==4.2.0.32
tqdm
json
codecs
gzip
glob
tabulate